pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pytrinamic"
authors = [
    { name="ADI Trinamic Software Team", email="tmc_info@trinamic.com" },
]
description = "TRINAMIC's Python Technology Access Package."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-can>=3",
    "canopen",
    "pyserial>=3",
    "IntelHex>=2.3",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/analogdevicesinc/pytrinamic"

[project.scripts]
tmclfwupload = "pytrinamic.cli.tmclfwupload:main"

[tool.setuptools.packages]
find = {}

[tool.setuptools.dynamic]
version = { attr = "pytrinamic.version.__version__" }